pr-review: also move YamlWriter
